Fixed type mismatch

+++ autoscaler/master/master.go
@@ -291,7 +291,7 @@ func (m *Master) reload() {
 
 func (m *Master) streamStats() {
 	m.statsdClientBuffer.FGauge("loadavg", m.currentLoadAvg)
-	m.statsdClientBuffer.Gauge("workers", len(m.droplets))
+	m.statsdClientBuffer.Gauge("workers", int64(len(m.droplets)))
 	fmt.Println("Streamed to statsd")
 	time.Sleep(time.Second * 5)
 }
